Moved password authorization from social directory
