+++ modules/user/server-ts/password/index.js
@@ -1,7 +1,7 @@
 import { AuthModule } from '@gqlapp/authentication-server-ts';
 import schema from './schema.graphql';
 import resolvers from './resolvers';
-import settings from '../../../../../settings';
+import settings from '../../../../settings';
 
 export default (settings.auth.password.enabled
   ? new AuthModule({ schema: [schema], createResolversFunc: [resolvers] })
