+++ modules/user/server-ts/social/index.js
@@ -3,7 +3,6 @@ import facebook, { facebookData } from './facebook';
 import github, { githubData } from './github';
 import google, { googleData } from './google';
 import linkedin, { linkedinData } from './linkedIn';
-import password from './password';
 
 const social = {
   ...linkedinData,
@@ -12,6 +11,6 @@ const social = {
   ...githubData
 };
 
-export default new AuthModule(facebook, github, google, linkedin, password, {
+export default new AuthModule(facebook, github, google, linkedin, {
   appContext: { social }
 });
